// src/app/api/model/user.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Plato } from './plato';
import { Direccion } from './direccion';


export interface User { 
    id?: string;
    nickname?: string;
    email?: string;
    iconUrl?: string;
    direcciones?: Set<Direccion>;
    platosFavoritos?: Set<Plato>;
}